+++ common/src/main/java/immersive_armors/Main.java
@@ -1,6 +1,7 @@
 package immersive_armors;
 
 import immersive_armors.network.NetworkManager;
+import net.minecraft.util.Identifier;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,4 +13,8 @@ public final class Main {
     public static boolean FORGE = false;
 
     public static boolean ENABLE_EFFECTS = Config.getInstance().enableEffects;
+
+    public static Identifier locate(String path) {
+        return new Identifier(MOD_ID, path);
+    }
 }